Reorganise plot server code
Split types in separate model classes, remove use of dataclasses, update
typing annotation and README
