+++ .github/workflows/python-package.yml
@@ -43,7 +43,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install flake8 pytest
-        if [ -f api/requirements.txt ]; then pip install -r api/requirements.txt; fi
+        if [ -f davidia/requirements.txt ]; then pip install -r davidia/requirements.txt; fi
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
@@ -53,4 +53,4 @@ jobs:
     - name: Test with pytest
       run: |
         if [[ ! -e build ]]; then mkdir build; fi
-        pytest -vvv api/tests/
+        PYTHONPATH=. pytest -vvv davidia/tests/
