+++ .pydevproject
@@ -6,6 +6,6 @@
     <pydev_property name="org.python.pydev.PYTHON_PROJECT_VERSION">python interpreter</pydev_property>
     
     <pydev_pathproperty name="org.python.pydev.PROJECT_SOURCE_PATH">
-        <path>/${PROJECT_DIR_NAME}/api</path>
+        <path>/${PROJECT_DIR_NAME}/</path>
     </pydev_pathproperty>
 </pydev_project>
